requirements/base.txt
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
python-decouple==3.8
psycopg2==2.9.9
requests==2.31.0
